chore: simplify logic
